+++ sources/ft_calloc.c
@@ -6,13 +6,12 @@
 /*   By: afuchs <afuchs@student.42mulhouse.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/23 17:21:39 by afuchs            #+#    #+#             */
-/*   Updated: 2022/02/23 18:13:53 by afuchs           ###   ########.fr       */
+/*   Updated: 2022/02/24 11:42:28 by afuchs           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
-#include <stdlib.h>
 #include "libft.h"
 
-void	*ft_calloc(unsigned int count, unsigned int size)
+void	*ft_calloc(size_t count, size_t size)
 {
 	void	*mem;
