+++ sources/ft_memchr.c
@@ -6,11 +6,12 @@
 /*   By: afuchs <afuchs@student.42mulhouse.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/23 12:58:14 by afuchs            #+#    #+#             */
-/*   Updated: 2022/02/23 13:19:56 by afuchs           ###   ########.fr       */
+/*   Updated: 2022/02/24 11:31:25 by afuchs           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-void	*ft_memchr(const void *s, int c, unsigned int n)
+void	*ft_memchr(const void *s, int c, size_t n)
 {
 	unsigned int	i;
