+++ sources/ft_memset.c
@@ -6,11 +6,12 @@
 /*   By: afuchs <afuchs@student.42mulhouse.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/22 12:01:55 by afuchs            #+#    #+#             */
-/*   Updated: 2022/02/23 17:45:59 by afuchs           ###   ########.fr       */
+/*   Updated: 2022/02/24 11:20:44 by afuchs           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-void	*ft_memset(void *b, int c, unsigned int len)
+void	*ft_memset(void *b, int c, size_t len)
 {
 	unsigned int	i;
