+++ sources/ft_strnstr.c
@@ -6,11 +6,12 @@
 /*   By: afuchs <afuchs@student.42mulhouse.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/02/23 14:03:52 by afuchs            #+#    #+#             */
-/*   Updated: 2022/02/23 14:55:19 by afuchs           ###   ########.fr       */
+/*   Updated: 2022/02/24 11:36:25 by afuchs           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+#include "libft.h"
 
-char	*ft_strnstr(const char *haystack, const char *needle, unsigned int len)
+char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
 	unsigned int	i;
 	unsigned int	j;
